+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Release 0.7.2.0
 ---------------
 
 * Removed dependency on `open-browser` package.
+* Starting with this version, `pandoc` 2.9 is **no longer supported** due to a breaking API change in `pandoc` 2.10.
 
 Release 0.7.1.0
 ---------------